Add a rOpenGov badge to your README file

Add an rOpenGov badge pointing to https://ropengov.org/ on your README file:

Usage

rog_badge_ropengov(install = TRUE)

Arguments

install

TRUE/FALSE. On TRUE the badge would be installed on your README.md/.Rmd. On FALSE a message with the Rmarkdown code would be displayed.

Value

A markdown badge or a message with the markdown code to be pasted.

See also

Examples

rog_badge_ropengov(install = FALSE)
#> Badge URL is 
#> [![rOG-badge](https://ropengov.github.io/rogtemplate/reference/figures/ropengov-badge.svg)](https://ropengov.org/)